backend/pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "invest-backend"
version = "0.1.0"
description = "Investment Dashboard Backend API"
authors = [{name = "Investment Team", email = "team@invest.com"}]
dependencies = [
    "fastapi==0.110.0",
    "uvicorn[standard]==0.29.0",
    "SQLAlchemy==2.0.23",
    "asyncpg==0.29.0",
    "psycopg2-binary==2.9.9",
    "pydantic[email]==2.3.0",
    "email-validator==2.1.0.post1",
    "alembic==1.12.1",
    "python-jose==3.3.0",
    "passlib[bcrypt]==1.7.4",
    "bcrypt==4.0.1",
    "celery==5.3.4",
    "redis==5.0.0",
    "httpx==0.27.0",
    "pytest==7.4.0",
    "pytest-asyncio==0.21.1",
    "ruff==0.1.0",
    "black==23.7.0",
    "openpyxl==3.1.2",
    "aiosqlite==0.19.0",
    "python-multipart==0.0.9",
]
requires-python = ">=3.11"

[tool.black]
line-length = 100
target-version = ["py311"]

[tool.ruff]
line-length = 100
select = ["E","F","I","UP"]
ignore = ["E501"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
asyncio_mode = "auto"
